Implement Spring Security

+++ src/main/java/com/myprojects/projects/airbnb/service/impl/HotelServiceImpl.java
@@ -5,7 +5,9 @@ import com.myprojects.projects.airbnb.dto.HotelInfoDto;
 import com.myprojects.projects.airbnb.dto.RoomDto;
 import com.myprojects.projects.airbnb.entity.Hotel;
 import com.myprojects.projects.airbnb.entity.Room;
+import com.myprojects.projects.airbnb.entity.User;
 import com.myprojects.projects.airbnb.exception.ResourceNotFoundException;
+import com.myprojects.projects.airbnb.exception.UnAuthorizedException;
 import com.myprojects.projects.airbnb.repository.HotelRepository;
 import com.myprojects.projects.airbnb.repository.RoomRepository;
 import com.myprojects.projects.airbnb.service.HotelService;
@@ -14,6 +16,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 
@@ -36,6 +39,8 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating new hotel: {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         Hotel savedhotel = hotelRepository.save(hotel);
         log.info("Created a new Hotel with ID: {}", savedhotel.getId());
         return modelMapper.map(savedhotel, HotelDto.class);
@@ -45,6 +50,11 @@ public class HotelServiceImpl implements HotelService {
     public HotelDto getHotelById(Long id) {
         log.info("Getting the hotel with ID: {}", id);
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID:" + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            log.error("Unauthorized access attempt by user: {}", user.getId());
+            throw new UnAuthorizedException("This User does not own this hotel with ID: " + id);
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -52,6 +62,11 @@ public class HotelServiceImpl implements HotelService {
     public HotelDto updateHotelById(Long id, HotelDto hotelDto) {
         log.info("Updating the hotel with ID: {}", id);
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID:" + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            log.error("Unauthorized access attempt by user: {}", user.getId());
+            throw new UnAuthorizedException("This User does not own this hotel with ID: " + id);
+        }
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -62,7 +77,11 @@ public class HotelServiceImpl implements HotelService {
     @Transactional
     public void deleteHotelById(Long id) {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID:" + id));
-
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            log.error("Unauthorized access attempt by user: {}", user.getId());
+            throw new UnAuthorizedException("This User does not own this hotel with ID: " + id);
+        }
         for (Room room : hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -76,7 +95,14 @@ public class HotelServiceImpl implements HotelService {
         log.info("Activating the hotel with ID: {}", id);
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new RuntimeException("Hotel not found with ID:" + id));
         hotel.setActive(true);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            log.error("Unauthorized access attempt by user: {}", user.getId());
+            throw new UnAuthorizedException("This User does not own this hotel with ID: " + id);
+        }
         //Only once
+
         for (Room room : hotel.getRooms()) {
             inventoryService.initializeRoomForAYear(room);
         }
@@ -90,7 +116,7 @@ public class HotelServiceImpl implements HotelService {
         return hotel.stream().map((receivedHotel) -> modelMapper.map(receivedHotel, HotelDto.class)).collect(Collectors.toList());
 
     }
-
+    //public method to get hotel info by ID, including rooms
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         Hotel hotel=hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
